Step #4
added #ifdef _DEBUG:, which is a preprocessor directive that checks if the code is being compiled in debug mode. If it is, it executes the code block within the #ifdef and #endif.

added print student information inside the debug block, I added a loop to print the firstName and lastName of each student.